% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenariovalidationcriteria-package.R
\docType{package}
\name{scenariovalidationcriteria-package}
\alias{scenariovalidationcriteria}
\alias{scenariovalidationcriteria-package}
\title{Scenario Vetting Criteria R-package}
\description{
Common definitions of vetting criteria for IAM scenarios
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/IAMconsortium/scenario-validation-criteria/}
}

}
\author{
Philipp C. Verpoort \href{mailto:philipp.verpoort@pik-potsdam.de}{philipp.verpoort@pik-potsdam.de}
}
\keyword{Assessment}
\keyword{Climate}
\keyword{Criteria}
\keyword{Emissions}
\keyword{Energy}
\keyword{Integrated}
\keyword{Land-use}
\keyword{Modelling}
\keyword{Pathways}
\keyword{Scenario}
\keyword{Validation}
\keyword{Vetting}
